Move gaussian_blur to losses (fixes plaidml race condition)
